+++ changelog.md
@@ -6,6 +6,7 @@ Features
 * Respond to `-h` alone with the helpdoc.
 * Allow `--hostname` as an alias for `--host`.
 * Suggest tables with foreign key relationships for JOIN and ON (#975)
+* Deprecate `$DSN` environment variable in favor of `$MYSQL_DSN`.
 
 
 Bug Fixes